+++ src/prefect/tasks/airtable/__init__.py
@@ -1,7 +1,5 @@
 """
 A collection of tasks for interacting with Airtable.
-
-The default Airtable credential secret name is `"AIRTABLE_API_KEY"`
 """
 try:
     from prefect.tasks.airtable.airtable import WriteAirtableRow