+++ src/prefect/tasks/github/issues.py
@@ -1,5 +1,6 @@
 import json
 from typing import Any, List
+import warnings
 
 import requests
 
@@ -19,8 +20,7 @@ class OpenGitHubIssue(Task):
         - body (str, optional): the contents of the issue; can also be provided to the `run` method
         - labels (List[str], optional): a list of labels to apply to the newly opened issues; can also
             be provided to the `run` method
-        - token_secret (str, optional): the name of the Prefect Secret containing your GitHub Access Token;
-            defaults to "GITHUB_ACCESS_TOKEN"
+        - token_secret (str, optional, DEPRECATED): the name of the Prefect Secret containing your GitHub Access Token
         - **kwargs (Any, optional): additional keyword arguments to pass to the standard Task init method
     """
 
@@ -30,13 +30,19 @@ class OpenGitHubIssue(Task):
         title: str = None,
         body: str = None,
         labels: List[str] = None,
-        token_secret: str = "GITHUB_ACCESS_TOKEN",
+        token_secret: str = None,
         **kwargs: Any
     ):
         self.repo = repo
         self.title = title
         self.body = body
         self.labels = labels or []
+        if token_secret is not None:
+            warnings.warn(
+                "The `token` argument is deprecated. Use a `Secret` task "
+                "to pass the credentials value at runtime instead.",
+                UserWarning,
+            )
         self.token_secret = token_secret
         super().__init__(**kwargs)
 
@@ -47,6 +53,7 @@ class OpenGitHubIssue(Task):
         title: str = None,
         body: str = None,
         labels: List[str] = None,
+        token: str = None,
     ) -> None:
         """
         Run method for this Task. Invoked by calling this Task after initialization within a Flow context,
@@ -57,8 +64,9 @@ class OpenGitHubIssue(Task):
                 form `organization/repo_name`; defaults to the one provided at initialization
             - title (str, optional): the title of the issue to create; defaults to the one provided at initialization
             - body (str, optional): the contents of the issue; defaults to the one provided at initialization
-            - labels (List[str], optional): a list of labels to apply to the newly opened issues; defaults to the ones provided
-                at initialization
+            - labels (List[str], optional): a list of labels to apply to the newly opened issues; defaults to
+            the ones provided at initialization
+            - token (str): a GitHub API token
 
         Raises:
             - ValueError: if a `repo` was never provided
@@ -71,7 +79,14 @@ class OpenGitHubIssue(Task):
             raise ValueError("A GitHub repository must be provided.")
 
         ## prepare the request
-        token = Secret(self.token_secret).get()
+        if token is None:
+            warnings.warn(
+                "The `token` argument is deprecated. Use a `Secret` task "
+                "to pass the credentials value at runtime instead.",
+                UserWarning,
+            )
+            token = Secret(self.token_secret).get()
+
         url = "https://api.github.com/repos/{}/issues".format(repo)
         headers = {
             "AUTHORIZATION": "token {}".format(token),