+++ src/prefect/tasks/twitter/__init__.py
@@ -1,7 +1,5 @@
 """
 Tasks for interacting with Twitter.
-
-The default Twitter credential secret name is `"TWITTER_API_CREDENTIALS"`, and should be a JSON document with four keys: `"api_key"`, `"api_secret"`, `"access_token"`, and `"access_token_secret"`
 """
 try:
     from prefect.tasks.twitter.twitter import LoadTweetReplies