+++ tests/tasks/github/test_issues.py
@@ -14,7 +14,6 @@ class TestOpenGithubIssueInitialization:
         assert task.title is None
         assert task.body is None
         assert task.labels == []
-        assert task.token_secret == "GITHUB_ACCESS_TOKEN"
 
     def test_additional_kwargs_passed_upstream(self):
         task = OpenGitHubIssue(name="test-task", checkpoint=True, tags=["bob"])
@@ -22,9 +21,7 @@ class TestOpenGithubIssueInitialization:
         assert task.checkpoint is True
         assert task.tags == {"bob"}
 
-    @pytest.mark.parametrize(
-        "attr", ["repo", "body", "title", "labels", "token_secret"]
-    )
+    @pytest.mark.parametrize("attr", ["repo", "body", "title", "labels"])
     def test_initializes_attr_from_kwargs(self, attr):
         task = OpenGitHubIssue(**{attr: "my-value"})
         assert getattr(task, attr) == "my-value"
@@ -37,25 +34,12 @@ class TestOpenGithubIssueInitialization:
 
 class TestCredentialsandProjects:
     def test_creds_are_pulled_from_secret_at_runtime(self, monkeypatch):
-        task = OpenGitHubIssue()
+        task = OpenGitHubIssue(token_secret="GITHUB_ACCESS_TOKEN")
 
         req = MagicMock()
         monkeypatch.setattr("prefect.tasks.github.issues.requests", req)
 
-        with set_temporary_config({"cloud.use_local_secrets": True}):
-            with prefect.context(secrets=dict(GITHUB_ACCESS_TOKEN={"key": 42})):
-                task.run(repo="org/repo")
-
-        assert req.post.call_args[1]["headers"]["AUTHORIZATION"] == "token {'key': 42}"
-
-    def test_creds_secret_can_be_overwritten(self, monkeypatch):
-        task = OpenGitHubIssue(token_secret="MY_SECRET")
-
-        req = MagicMock()
-        monkeypatch.setattr("prefect.tasks.github.issues.requests", req)
-
-        with set_temporary_config({"cloud.use_local_secrets": True}):
-            with prefect.context(secrets=dict(MY_SECRET={"key": 42})):
-                task.run(repo="org/repo")
+        with prefect.context(secrets=dict(GITHUB_ACCESS_TOKEN={"key": 42})):
+            task.run(repo="org/repo")
 
         assert req.post.call_args[1]["headers"]["AUTHORIZATION"] == "token {'key': 42}"