+++ tests/tasks/github/test_repos.py
@@ -12,7 +12,6 @@ class TestGetRepoInfoInitialization:
         task = GetRepoInfo()
         assert task.repo is None
         assert task.info_keys == []
-        assert task.token_secret == "GITHUB_ACCESS_TOKEN"
 
     def test_additional_kwargs_passed_upstream(self):
         task = GetRepoInfo(name="test-task", checkpoint=True, tags=["bob"])
@@ -20,7 +19,7 @@ class TestGetRepoInfoInitialization:
         assert task.checkpoint is True
         assert task.tags == {"bob"}
 
-    @pytest.mark.parametrize("attr", ["repo", "info_keys", "token_secret"])
+    @pytest.mark.parametrize("attr", ["repo", "info_keys",])
     def test_initializes_attr_from_kwargs(self, attr):
         task = GetRepoInfo(**{attr: "my-value"})
         assert getattr(task, attr) == "my-value"
@@ -37,7 +36,6 @@ class TestCreateBranchInitialization:
         assert task.repo is None
         assert task.base == "master"
         assert task.branch_name is None
-        assert task.token_secret == "GITHUB_ACCESS_TOKEN"
 
     def test_additional_kwargs_passed_upstream(self):
         task = CreateBranch(name="test-task", checkpoint=True, tags=["bob"])
@@ -45,7 +43,7 @@ class TestCreateBranchInitialization:
         assert task.checkpoint is True
         assert task.tags == {"bob"}
 
-    @pytest.mark.parametrize("attr", ["repo", "base", "branch_name", "token_secret"])
+    @pytest.mark.parametrize("attr", ["repo", "base", "branch_name",])
     def test_initializes_attr_from_kwargs(self, attr):
         task = CreateBranch(**{attr: "my-value"})
         assert getattr(task, attr) == "my-value"
@@ -61,54 +59,28 @@ class TestCreateBranchInitialization:
             task.run()
 
 
-class TestCredentials:
+# deprecated
+class TestCredentialsSecret:
     def test_creds_are_pulled_from_secret_at_runtime_repo_info(self, monkeypatch):
-        task = GetRepoInfo()
+        task = GetRepoInfo(token_secret="GITHUB_ACCESS_TOKEN")
 
         req = MagicMock()
         monkeypatch.setattr("prefect.tasks.github.repos.requests", req)
 
-        with set_temporary_config({"cloud.use_local_secrets": True}):
-            with prefect.context(secrets=dict(GITHUB_ACCESS_TOKEN={"key": 42})):
-                task.run(repo="org/repo")
+        with prefect.context(secrets=dict(GITHUB_ACCESS_TOKEN={"key": 42})):
+            task.run(repo="org/repo")
 
         assert req.get.call_args[1]["headers"]["AUTHORIZATION"] == "token {'key': 42}"
 
     def test_creds_are_pulled_from_secret_at_runtime_create_branch(self, monkeypatch):
-        task = CreateBranch()
+        task = CreateBranch(token_secret="GITHUB_ACCESS_TOKEN")
 
         req = MagicMock()
         monkeypatch.setattr("prefect.tasks.github.repos.requests", req)
 
-        with set_temporary_config({"cloud.use_local_secrets": True}):
-            with prefect.context(secrets=dict(GITHUB_ACCESS_TOKEN={"key": 42})):
-                with pytest.raises(ValueError, match="not found"):
-                    task.run(repo="org/repo", branch_name="new")
-
-        assert req.get.call_args[1]["headers"]["AUTHORIZATION"] == "token {'key': 42}"
-
-    def test_creds_secret_can_be_overwritten_repo_info(self, monkeypatch):
-        task = GetRepoInfo(token_secret="MY_SECRET")
-
-        req = MagicMock()
-        monkeypatch.setattr("prefect.tasks.github.repos.requests", req)
-
-        with set_temporary_config({"cloud.use_local_secrets": True}):
-            with prefect.context(secrets=dict(MY_SECRET={"key": 42})):
-                task.run(repo="org/repo")
-
-        assert req.get.call_args[1]["headers"]["AUTHORIZATION"] == "token {'key': 42}"
-
-    def test_creds_secret_can_be_overwritten_create_branch(self, monkeypatch):
-        task = CreateBranch(token_secret="MY_SECRET")
-
-        req = MagicMock()
-        monkeypatch.setattr("prefect.tasks.github.repos.requests", req)
-
-        with set_temporary_config({"cloud.use_local_secrets": True}):
-            with prefect.context(secrets=dict(MY_SECRET={"key": 42})):
-                with pytest.raises(ValueError):
-                    task.run(repo="org/repo", branch_name="new")
+        with prefect.context(secrets=dict(GITHUB_ACCESS_TOKEN={"key": 42})):
+            with pytest.raises(ValueError, match="not found"):
+                task.run(repo="org/repo", branch_name="new")
 
         assert req.get.call_args[1]["headers"]["AUTHORIZATION"] == "token {'key': 42}"
 
@@ -122,9 +94,7 @@ def test_base_name_is_filtered_for(monkeypatch):
     )
     monkeypatch.setattr("prefect.tasks.github.repos.requests", req)
 
-    with set_temporary_config({"cloud.use_local_secrets": True}):
-        with prefect.context(secrets=dict(GITHUB_ACCESS_TOKEN={"key": 42})):
-            task.run(repo="org/repo")
+    task.run(repo="org/repo", token={"key": 42})
 
     assert req.post.call_args[1]["json"] == {
         "ref": "refs/heads/NEWBRANCH",